+++ core/includes/binder/memory/stringIntern.h
@@ -5,28 +5,23 @@
 namespace binder::memory {
 
 class StringIntern {
-public:
+ public:
   StringIntern(int bucketCount) : m_values(bucketCount) {}
 
   const char *intern(const char *string, bool copy = true) {
-    int len = strlen(string);
+    int len = static_cast<int>(strlen(string));
     return intern(string, len, copy);
   }
   const char *intern(const char *string, int len, bool copy = true) {
     const char *toReturn;
 
-    //assert(len<128);
-    //char temp[128];
-    //memcpy(temp,string,len+1),
-    //temp[len] = '\0';
-    
-    bool result = m_values.get(string,len, toReturn);
+    bool result = m_values.get(string, len, toReturn);
     if (result) {
       // string is present we intern it
       return toReturn;
     } else {
       // we need to copy the string and insert it
-      char *toInsert = (char*)string;
+      char *toInsert = (char *)string;
 
       if (copy) {
         toInsert = new char[len + 1];
@@ -38,7 +33,7 @@ public:
     }
   }
 
-private:
+ private:
   memory::HashMap<const char *, const char *, hashString32> m_values;
 };
-} // namespace binder::memory
+}  // namespace binder::memory
